refactor: require calibrated TOC offset and share hierarchy collapse
Drop arithmetic offset fallback and legacy page_locate budget so shard
planning only uses VLM-calibrated offsets, and reuse one collapse helper
across calibration and page_memory.
