+++ apps/worker/.env.example
@@ -141,8 +141,6 @@ PARSE_AGENT_CALIBRATION_MIN_BUDGET=12000
 PARSE_AGENT_CALIBRATION_CAP=40000
 PARSE_AGENT_PAGE_TAGGING_MIN_BUDGET=0
 PARSE_AGENT_PAGE_TAGGING_CAP=0
-PARSE_AGENT_PAGE_LOCATE_MIN_BUDGET=0
-PARSE_AGENT_PAGE_LOCATE_CAP=0
 
 # Parser row schema. `entities` (JSON typed entities, §4.4) and `asset_title`
 # (asset caption/label, §4.5) are additive trailing columns; the parser layer
