+++ apps/worker/app/services/document_agent/budget.py
@@ -12,7 +12,6 @@ BudgetStage = Literal[
     "coarse_planner",
     "structural_react",
     "calibration",
-    "page_locate",
     "page_tagging",
 ]
 
@@ -174,10 +173,6 @@ class BudgetTracker:
             for stage, envelope in self._visual_stage_envelopes.items():
                 usage = self._visual_stage_usage.get(stage, StageUsage())
                 remaining_guarantee = max(envelope.min_guarantee - usage.committed, 0)
-                if stage == "page_locate":
-                    # Legacy uncapped locate stage: keep child uncapped.
-                    child_envelopes[stage] = StageEnvelope(min_guarantee=0, cap=0)
-                    continue
                 child_envelopes[stage] = StageEnvelope(
                     min_guarantee=int(remaining_guarantee * ratio),
                     cap=int(envelope.cap * ratio) if envelope.cap is not None else None,
